mychat/requirements.txt
Django>=4.0,<5.0
channels>=3.0,<4.0
channels_redis>=3.0,<4.0
daphne>=3.0,<4.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.12,<4.0
redis>=4.0,<5.0   # Optional, if needed for Redis client
